Add displayActionsInToolbar and actionsPosition
